irq: timings: Fix next_prediction type to s64

+++ include/linux/irq.h
@@ -822,9 +822,9 @@ static inline void irq_gc_unlock(struct irq_chip_generic *gc) { }
 #endif
 
 #ifdef CONFIG_IRQ_TIMINGS
-extern u32 irqt_get_next_prediction(int cpu);
+extern s64 irqt_get_next_prediction(int cpu);
 #else
-static inline u32 irqt_get_next_prediction(int cpu) { return 0; }
+static inline s64 irqt_get_next_prediction(int cpu) { return 0; }
 #endif
 
 #endif /* _LINUX_IRQ_H */
